requirements.txt
opencv-python
boto3
fastapi
uvicorn
python-multipart
uvicorn[standard]
PyMuPDF
openai
faiss-cpu
numpy
python-dotenv
newberryai
pydantic
requests
pandas
mysql-connector-python